feat(store): スポット操作の実装 (#2503)

+++ api/internal/store/database/database.go
@@ -495,7 +495,7 @@ type ListSpotsParams struct {
 type ListSpotsByGeolocationParams struct {
 	Longitude float64
 	Latitude  float64
-	Radius    int
+	Radius    int64
 }
 
 type UpdateSpotParams struct {
